reorganize repo utils
